1. Introduce ONNX conversion as an extention to MLContext 2. Address minor comments
Remove two best friends
